rm SlashInsistingRoutingRule & NoTrailingSlashError
I had a clever workaround to respond with 404 errors in case
of missing trailing slashes. The default strict_slashes=True
behavior is to redirect to the one with the slash, and it
gives na "Redirecting..." response, which is unwanted as an api
strict_slashes = False behaivor is to just accept the url without
the trailing slash, which is also not what I wanted
So I had overwritten the default Werkzeug Rule class for
the behavior I wanted: raise 404 with a helpful error message
about the missing trailing slash.
Werkzeug code has changed and now Rule classes do not have
the matching method, it's in the MapAdaptern class that a Map
creates via its bind method. Instead of overwriting a MapAdapter
and a new Map class using it to set as Flask app's default
url_map_class attribute, I said fudge it
and removed the whole thing, went with strict_slashes=False
Shame on Werkzeug and Flask for not having a good way to
implement the simple & reasonable api behaivor I want without
having to mess with the innards of Werkzeug.